refactor: centralize `Status` and `StatusDB` in `status` model, streamline `list_projects` with project status integration
